# services/wallet-service/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "Core Wallet API",
        "termsOfService": "https://docs.wallet.example/terms",
        "contact": {
            "name": "Core Wallet Team",
            "email": "hoalh2@hdbank.com.vn"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/accounts": {
            "post": {
                "description": "Open a zero-balance wallet for a client (count-limited per currency).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Open a wallet",
                "parameters": [
                    {
                        "description": "Open account request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OpenAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountOpenResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation (e.g. account count limit)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/accounts/{acct_no}": {
            "get": {
                "description": "Account profile (no client PII).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            },
            "patch": {
                "description": "Block / close / re-activate an account (status A | B | C).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Update account status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAccountStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation (e.g. illegal transition)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/accounts/{acct_no}/balance": {
            "get": {
                "description": "Without as_of_date → customer realtime balance (§9.3.1). With as_of_date → historical end-of-day snapshot (§9.3.3, returns BalanceAsOfResponse).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Historical EOD snapshot date (YYYY-MM-DD)",
                        "name": "as_of_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Realtime balance (or BalanceAsOfResponse when as_of_date is set)",
                        "schema": {
                            "$ref": "#/definitions/dto.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/v1/ops/accounts/balance/batch": {
            "post": {
                "description": "Query up to 100 account balances at once. Unknown accounts are returned per-item with error=ACCT_NOT_FOUND (§9.3.4).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Batch balance query",
                "parameters": [
                    {
                        "description": "Batch balance request (max 100 accounts)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BalanceBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BalanceBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/ops/accounts/{acct_no}/balance": {
            "get": {
                "description": "Ops/internal full balance view incl. ledger/calc/restrained breakdown and active restraints (§9.3.2).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Get account balance (ops full view)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BalanceOpsResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AccountOpenResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "acct_type": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "client_no": {
                    "type": "string"
                }
            }
        },
        "dto.AccountResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_open_date": {
                    "type": "string"
                },
                "acct_role": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "acct_type": {
                    "type": "string"
                },
                "actual_bal": {
                    "type": "string"
                },
                "available_bal": {
                    "description": "calc_bal = actual - restrained",
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "client_no": {
                    "type": "string"
                },
                "cr_blocked": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "last_tran_date": {
                    "type": "string"
                },
                "prev_day_bal": {
                    "type": "string"
                },
                "restrained_amt": {
                    "type": "string"
                },
                "restraint_present": {
                    "type": "string"
                },
                "shard_index": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "dto.AccountStatusResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "dto.BalanceBatchItem": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "actual_bal": {
                    "type": "string"
                },
                "available_bal": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "restrained_amt": {
                    "type": "string"
                }
            }
        },
        "dto.BalanceBatchRequest": {
            "type": "object",
            "required": [
                "acct_nos"
            ],
            "properties": {
                "acct_nos": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.BalanceBatchResponse": {
            "type": "object",
            "properties": {
                "as_of": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BalanceBatchItem"
                    }
                }
            }
        },
        "dto.BalanceOpsResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "active_restraints": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "actual_bal": {
                    "type": "string"
                },
                "as_of": {
                    "type": "string"
                },
                "available_bal": {
                    "type": "string"
                },
                "calc_bal": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "client_no": {
                    "type": "string"
                },
                "cr_blocked": {
                    "type": "string"
                },
                "last_tran_date": {
                    "type": "string"
                },
                "ledger_bal": {
                    "type": "string"
                },
                "previous_day_bal": {
                    "type": "string"
                },
                "restrained_amt": {
                    "type": "string"
                },
                "restraint_present": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "dto.BalanceResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "actual_bal": {
                    "type": "string"
                },
                "as_of": {
                    "type": "string"
                },
                "available_bal": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "last_tran_date": {
                    "type": "string"
                },
                "masked": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "restrained_amt": {
                    "type": "string"
                }
            }
        },
...
        "dto.FieldError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "dto.OpenAccountRequest": {
            "type": "object",
            "required": [
                "acct_type",
                "client_no"
            ],
            "properties": {
                "acct_type": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string",
                    "maxLength": 4
                },
                "client_no": {
                    "type": "string",
                    "maxLength": 48
                }
            }
        },
        "dto.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "RFC7807: human-readable detail (dynamic context)",
                    "type": "string",
                    "example": "available 50000 \u003c required 100000"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "errorCode": {
                    "description": "stable business error code (canonical contract)",
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "errorMessage": {
                    "description": "stable human message for this code (i18n source, safe for end-user)",
                    "type": "string",
                    "example": "The account balance is not sufficient to cover this transaction"
                },
                "errors": {
                    "description": "field-level (Berlin Group / OBIE style)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FieldError"
                    }
                },
                "instance": {
                    "description": "RFC7807: request path",
                    "type": "string",
                    "example": "/v1/finance/transfer"
                },
                "internal_code": {
                    "description": "E#### for log/alert (§5)",
                    "type": "string",
                    "example": "E4022"
                },
                "iso20022_reason_code": {
                    "description": "ISO 20022 External Status Reason (§13.2)",
                    "type": "string",
                    "example": "AM04"
                },
                "retry": {
                    "$ref": "#/definitions/dto.RetryInfo"
                },
                "status": {
                    "description": "RFC7807: HTTP status code",
                    "type": "integer",
                    "example": 422
                },
                "timestamp": {
                    "description": "RFC 3339",
                    "type": "string",
                    "example": "2026-06-06T07:30:00Z"
                },
                "title": {
                    "description": "RFC7807: short human title",
                    "type": "string",
                    "example": "Insufficient funds"
                },
                "trace_id": {
                    "description": "= X-Request-Id",
                    "type": "string",
                    "example": "5f3b2c8e-1a2b-4c3d-9e8f-0a1b2c3d4e5f"
                },
                "transaction_status": {
                    "description": "pain.002 status (§13.3)",
                    "type": "string",
                    "example": "RJCT"
                },
                "type": {
                    "description": "RFC7807: URI to error doc",
                    "type": "string",
                    "example": "https://docs.wallet.example/errors/INSUFFICIENT_FUNDS"
                }
            }
        },
...
        "dto.RetryInfo": {
            "type": "object",
            "properties": {
                "after_ms": {
                    "type": "integer"
                },
                "retryable": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateAccountStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "A",
                        "B",
                        "C"
                    ]
                }
            }
        },
...
    },
...
}